fix bus errors on model construction and (hopefully) loading for big models

+++ csrc/learn.c
@@ -19,7 +19,7 @@ typedef long long int Int;
 float inplace_update(float* In, float* Out, 
 	Int M, Int T, double* z,
 	Int x,  
-	int32_t* path, int8_t* code, int64_t length,
+	int64_t* path, int8_t* code, int64_t length,
 	float* in_grad, float* out_grad, 
 	float lr, float sense_treshold) {
 
@@ -27,23 +27,23 @@ float inplace_update(float* In, float* Out,
 
 	float pr = 0;
 
-	for (int k = 0; k < T; ++k)
-		for (int i = 0; i < M; ++i)
+	for (Int k = 0; k < T; ++k)
+		for (Int i = 0; i < M; ++i)
 			in_grad[k*M + i] = 0;
 
-	for (int n = 0; n < length && code[n] != -1; ++n) {
+	for (Int n = 0; n < length && code[n] != -1; ++n) {
 		float* out = Out + (path[n]-1)*M;
 
-		for (int i = 0; i < M; ++i)
+		for (Int i = 0; i < M; ++i)
 			out_grad[i] = 0;
 
-		for (int k = 0; k < T; ++k) {
+		for (Int k = 0; k < T; ++k) {
 			if (z[k] < sense_treshold) continue;
 
 			float* in = in_offset(In, x, k, M, T);
 
 			float f = 0;
-			for (int i = 0; i < M; ++i)
+			for (Int i = 0; i < M; ++i)
 				f += in[i] * out[i];
 
 			pr += z[k] * logsigmoid(f * (1 - 2*code[n]));
@@ -51,20 +51,20 @@ float inplace_update(float* In, float* Out,
 			float d = 1 - code[n] - sigmoid(f);
 			float g = z[k] * lr * d;
 
-			for (int i = 0; i < M; ++i) {
+			for (Int i = 0; i < M; ++i) {
 				in_grad[k*M + i] += g * out[i];
 				out_grad[i]      += g * in[i];
 			}
 		}
 
-		for (int i = 0; i < M; ++i)
+		for (Int i = 0; i < M; ++i)
 			out[i] += out_grad[i];
 	}
 
-	for (int k = 0; k < T; ++k) {
+	for (Int k = 0; k < T; ++k) {
 		if (z[k] < sense_treshold) continue;
 		float* in = in_offset(In, x, k, M, T);
-		for (int i = 0; i < M; ++i)
+		for (Int i = 0; i < M; ++i)
 			in[i] += in_grad[k*M + i];
 	}
 
@@ -73,15 +73,15 @@ float inplace_update(float* In, float* Out,
 
 float skip_gram(float* In, float* Out, 
 	Int M, 
-	int32_t* path, int8_t* code, int length) {
+	int64_t* path, int8_t* code, Int length) {
 
 	float pr = 0;
 
-	for (int n = 0; n < length && code[n] != -1; ++n) {
+	for (Int n = 0; n < length && code[n] != -1; ++n) {
 		float* out = Out + (path[n]-1)*M;
 
 		float f = 0;
-		for (int i = 0; i < M; ++i)
+		for (Int i = 0; i < M; ++i)
 			f += In[i] * out[i];
 
 		pr += logsigmoid(f * (1 - 2*code[n]));
@@ -93,18 +93,18 @@ float skip_gram(float* In, float* Out,
 void update_z(float* In, float* Out, 
 	Int M, Int T, double* z,
 	Int x,  
-	int32_t* path, int8_t* code, int64_t length) {
+	int64_t* path, int8_t* code, int64_t length) {
 
 	--x;
 
-	for (int n = 0; n < length && code[n] != -1; ++n) {
+	for (Int n = 0; n < length && code[n] != -1; ++n) {
 		float* out = Out + (path[n]-1)*M;
 
-		for (int k = 0; k < T; ++k) {
+		for (Int k = 0; k < T; ++k) {
 			float* in = in_offset(In, x, k, M, T);
 
 			float f = 0;
-			for (int i = 0; i < M; ++i)
+			for (Int i = 0; i < M; ++i)
 				f += in[i] * out[i];
 
 			z[k] += logsigmoid(f * (1 - 2*code[n]));
